Start on module support and clean up parser
