--- parser/expression_bind.go
@@ -1,30 +1,0 @@
-package parser
-
-import (
-	"fmt"
-
-	"ghostlang.org/x/ghost/ast"
-)
-
-func (p *Parser) parseBindExpression(expression ast.Expression) ast.Expression {
-	switch node := expression.(type) {
-	case *ast.IdentifierLiteral:
-	default:
-		message := fmt.Sprintf("expected identifier expression on left but got %T (%+v)", node, expression)
-		p.errors = append(p.errors, message)
-		return nil
-	}
-
-	be := &ast.BindExpression{Token: p.currentToken, Left: expression}
-
-	p.nextToken()
-
-	be.Value = p.parseExpression(LOWEST)
-
-	if fl, ok := be.Value.(*ast.FunctionLiteral); ok {
-		identifier := be.Left.(*ast.IdentifierLiteral)
-		fl.Name = identifier.Value
-	}
-
-	return be
-}
