--- parser/expression_if.go
@@ -1,55 +1,0 @@
-package parser
-
-import (
-	"ghostlang.org/x/ghost/ast"
-	"ghostlang.org/x/ghost/token"
-)
-
-func (p *Parser) parseIfExpression() ast.Expression {
-	expression := &ast.IfExpression{Token: p.currentToken}
-
-	if !p.expectPeek(token.LPAREN) {
-		return nil
-	}
-
-	p.nextToken()
-
-	expression.Condition = p.parseExpression(LOWEST)
-
-	if !p.expectPeek(token.RPAREN) {
-		return nil
-	}
-
-	if !p.expectPeek(token.LBRACE) {
-		return nil
-	}
-
-	expression.Consequence = p.parseBlockStatement()
-
-	if p.peekTokenIs(token.ELSE) {
-		p.nextToken()
-
-		// else if
-		if p.peekTokenIs(token.IF) {
-			p.nextToken()
-
-			expression.Alternative = &ast.BlockStatement{
-				Statements: []ast.Statement{
-					&ast.ExpressionStatement{
-						Expression: p.parseIfExpression(),
-					},
-				},
-			}
-
-			return expression
-		}
-
-		if !p.expectPeek(token.LBRACE) {
-			return nil
-		}
-
-		expression.Alternative = p.parseBlockStatement()
-	}
-
-	return expression
-}
