--- parser/literal_string.go
@@ -1,7 +1,0 @@
-package parser
-
-import "ghostlang.org/x/ghost/ast"
-
-func (p *Parser) parseStringLiteral() ast.Expression {
-	return &ast.StringLiteral{Token: p.currentToken, Value: p.currentToken.Literal}
-}
