--- parser/statement_expression.go
@@ -1,18 +1,0 @@
-package parser
-
-import (
-	"ghostlang.org/x/ghost/ast"
-	"ghostlang.org/x/ghost/token"
-)
-
-func (p *Parser) parseExpressionStatement() *ast.ExpressionStatement {
-	statement := &ast.ExpressionStatement{Token: p.currentToken}
-
-	statement.Expression = p.parseExpression(LOWEST)
-
-	if p.peekTokenIs(token.SEMICOLON) {
-		p.nextToken()
-	}
-
-	return statement
-}
